Fix animation curve.
Animation curve setting wasn't propagating due to use of the
non-bit-masked version of the constant.

As a result, the behavior was EaseInEaseOut, despite being set to
EaseIn, and the code base provided a bad example of which constant to
use to set the curve externally.

See issue #13 https://github.com/cbess/AutoScrollLabel/issues/13